#105 test: 현재 방문한 여행지가 최근 클릭한 선호 여행지 리스트에 포함된다면 선호 여행지들의 랭크를 수정하지 않는지 테스트

+++ backend/src/main/java/moheng/trip/application/TripService.java
@@ -70,9 +70,11 @@ public class TripService {
             memberTrip.incrementVisitedCount();
         }
         else if(recommendTrips.size() >= RECOMMEND_TRIPS_SIZE) {
-            downAllRanks(recommendTrips);
-            deleteHighestPriorityRankRecommendTrip(member);
-            recommendTripRepository.save(new RecommendTrip(trip, member, LOWEST_PRIORITY_RANK));
+            if(!recommendTripRepository.existsByMemberAndTrip(member, trip)) {
+                downAllRanks(recommendTrips);
+                deleteHighestPriorityRankRecommendTrip(member);
+                recommendTripRepository.save(new RecommendTrip(trip, member, LOWEST_PRIORITY_RANK));
+            }
             final MemberTrip memberTrip = findOrCreateMemberTrip(member, trip);
             memberTrip.incrementVisitedCount();
         }
